requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
pytz==2024.1